renaming everything!
renaming 'tests' to 'qc'
