+++ src/toolbox/steps/custom/qc/impossible_speed_qc.py
@@ -17,7 +17,7 @@
 """QC test to identify impossible speeds in glider data."""
 
 #### Mandatory imports ####
-from toolbox.steps.base_test import BaseTest, register_qc, flag_cols
+from toolbox.steps.base_qc import BaseTest, register_qc, flag_cols
 
 #### Custom imports ####
 import matplotlib.pyplot as plt
@@ -28,7 +28,7 @@ import matplotlib
 
 
 @register_qc
-class impossible_speed_test(BaseTest):
+class impossible_speed_qc(BaseTest):
     """
     Target Variable: TIME, LATITUDE, LONGITUDE
     Flag Number: 4 (bad data)
@@ -36,7 +36,7 @@ class impossible_speed_test(BaseTest):
     Checks that the the gliders horizontal speed stays below 3m/s
     """
 
-    test_name = "impossible speed test"
+    test_name = "impossible speed qc"
     expected_parameters = {}
     required_variables = ["TIME", "LATITUDE", "LONGITUDE"]
     qc_outputs = ["TIME_QC", "LATITUDE_QC", "LONGITUDE_QC"]
