+++ src/toolbox/steps/custom/qc/position_on_land_qc.py
@@ -14,32 +14,34 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""QC test to identify impossible dates in TIME variable."""
+"""QC test that identifies glider positions not located on land and flags accordingly."""
 
 #### Mandatory imports ####
-from toolbox.steps.base_test import BaseTest, register_qc, flag_cols
+from toolbox.steps.base_qc import BaseTest, register_qc, flag_cols
 
 #### Custom imports ####
+from geodatasets import get_path
+import matplotlib.pyplot as plt
+import shapely as sh
 import polars as pl
 import xarray as xr
-from datetime import datetime
 import matplotlib
-import matplotlib.pyplot as plt
+import geopandas
 
 
 @register_qc
-class impossible_date_test(BaseTest):
+class position_on_land_qc(BaseTest):
     """
-    Target Variable: TIME
+    Target Variable: LATITUDE, LONGITUDE
     Flag Number: 4 (bad data)
-    Variables Flagged: TIME
-    Checks that the datetime of each point is valid.
+    Variables Flagged: LATITUDE, LONGITUDE
+    Checks that the measurement location is not on land.
     """
 
-    test_name = "impossible date test"
+    test_name = "position on land qc"
     expected_parameters = {}
-    required_variables = ["TIME"]
-    qc_outputs = ["TIME_QC"]
+    required_variables = ["LATITUDE", "LONGITUDE"]
+    qc_outputs = ["LATITUDE_QC", "LONGITUDE_QC"]
 
     def return_qc(self):
         # Convert to polars
@@ -47,21 +49,26 @@ class impossible_date_test(BaseTest):
             self.data[self.required_variables].to_dataframe(), nan_to_null=False
         )
 
-        # Check if any of the datetime stamps fall outside 1985 and the current datetime
-        # TODO: Add optional bounds via parameters (such as known deployment dates, for example)
+        # Concat the polygons into a MultiPolygon object
+        self.world = geopandas.read_file(get_path("naturalearth.land"))
+        land_polygons = sh.ops.unary_union(self.world.geometry)
+
+        # Check if lat, long coords fall within the area of the land polygons
         self.df = self.df.with_columns(
-            pl.when(pl.col("TIME").is_null())
-            .then(9)
-            .when(
-                (
-                    (pl.col("TIME") > datetime(1985, 1, 1))
-                    & (pl.col("TIME") < datetime.now())
+            pl.struct("LONGITUDE", "LATITUDE")
+            .map_batches(
+                lambda x: sh.contains_xy(
+                    land_polygons,
+                    x.struct.field("LONGITUDE").to_numpy(),
+                    x.struct.field("LATITUDE").to_numpy(),
                 )
+                * 4
             )
-            .then(1)
-            .otherwise(4)
-            .alias("TIME_QC")
+            .replace({0: 1})
+            .alias("LONGITUDE_QC")
         )
+        # Add the flags to LATITUDE as well.
+        self.df = self.df.with_columns(pl.col("LONGITUDE_QC").alias("LATITUDE_QC"))
 
         # Convert back to xarray
         flags = self.df.select(pl.col("^.*_QC$"))
@@ -76,26 +83,31 @@ class impossible_date_test(BaseTest):
 
     def plot_diagnostics(self):
         matplotlib.use("tkagg")
-        fig, ax = plt.subplots(figsize=(6, 4), dpi=200)
+        fig, ax = plt.subplots(figsize=(12, 8), dpi=200)
+
+        # Plot land boundaries
+        self.world.plot(ax=ax, facecolor="lightgray", edgecolor="black", alpha=0.3)
+
         for i in range(10):
             # Plot by flag number
-            plot_data = self.df.with_row_index().filter(pl.col("TIME_QC") == i)
+            plot_data = self.df.filter(pl.col("LATITUDE_QC") == i)
             if len(plot_data) == 0:
                 continue
 
             # Plot the data
             ax.plot(
-                plot_data["index"],
-                plot_data["TIME"],
+                plot_data["LONGITUDE"],
+                plot_data["LATITUDE"],
                 c=flag_cols[i],
                 ls="",
                 marker="o",
                 label=f"{i}",
             )
+
         ax.set(
-            title="Impossible Date Test",
-            xlabel="Index",
-            ylabel="TIME",
+            xlabel="Longitude",
+            ylabel="Latitude",
+            title="Position On Land Test",
         )
         ax.legend(title="Flags", loc="upper right")
         fig.tight_layout()
