+++ tests/test_impossible_date_qc.py
@@ -4,7 +4,7 @@ import polars as pl
 import numpy as np
 import pandas as pd
 from datetime import datetime
-from toolbox.steps.custom.qc.impossible_date_test import impossible_date_test
+from toolbox.steps.custom.qc.impossible_date_qc import impossible_date_test
 
 # Helper function to create a test xarray Dataset
 def create_test_dataset(times):
